requirement.txt
Flask==2.3.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0